lib/pegasus: rename module to SRU
We are switching from Pegasus (Aleph) to Alma, so the name should be
more generic.

* move SRU host to secrets.yml

* move other variables to config/sru.yml

* better error messages

* traject: update #extract_work_type for Alma
