+++ config/application.yml
@@ -3,10 +3,10 @@ defaults: &defaults
   download_root: <%= ENV['ADRL_DOWNLOAD_ROOT'] || "#{Rails.root}/tmp/download_root" %>
   ezid_host: ezid.cdlib.org
   ezid_port: 443
+  ffmpeg_path: <%= ENV['FFMPEG_PATH'] || "ffmpeg" %>
   internal_local_vocab_root: http://<%= Rails.application.config.host_name %>
   minter_state: <%= ENV['ADRL_MINTER_STATE'] || "#{Rails.root}/tmp/minter-state" %>
   uploads_dir: <%= ENV['ADRL_UPLOADS'] || "#{Rails.root}/tmp/uploads" %>
-  ffmpeg_path: <%= ENV['FFMPEG_PATH'] || "ffmpeg" %>
 
 development:
   <<: *defaults
