+++ doc/ingesting.md
@@ -170,8 +170,8 @@ thesis, an XML file containing the metadata for the ETD, and
 (optionally) any supplemental files the author provided to ProQuest.
 
 We don’t ingest an ETD using the ProQuest metadata directly; instead
-we use the XML file to find the ETD in Aleph; then ingest it using the
-MARC metadata from Aleph.
+we use the XML file to find the ETD with the SRU API; then ingest it using the
+MARC metadata from Alma.
 
 What happens when you run `bin/ingest -f etd /path/to/etds/etdadmin_upload*`, then, is this:
 
@@ -192,9 +192,10 @@ What happens when you run `bin/ingest -f etd /path/to/etds/etdadmin_upload*`, th
 
 
 2. Next, `bin/ingest` passes the path to the XML file of each ETD to
-   {Importer::ETDParser.parse_file}, which parses the XML and queries Aleph,
-   returning a single string of MARC containing the metadata for all
-   the ETDs matched by `/path/to/etds/etdadmin_upload*`.
+   {Importer::ETDParser.parse_file}, which parses the XML and queries
+   the SRU API, returning a single string of MARC containing the
+   metadata for all the ETDs matched by
+   `/path/to/etds/etdadmin_upload*`.
 
 3. The MARC strings are passed (via `StringIO`) to `MARC::XMLReader`,
    then collected in an array.
@@ -239,9 +240,7 @@ instead of a single `marc` option.
 
 Unlike ETDs, for which the MARC is generated on the fly, currently
 ingesting cylinders requires a pre-existing MARCXML file; generated by
-querying the Aleph API (here’s how you might do it in PHP:
-<https://github.library.ucsb.edu/gist/ssmith/df34eb95bfc0a225a9f6>).
-Once the metadata file(s) have been generated,
+querying the SRU API.  Once the metadata file(s) have been generated,
 `bin/ingest -f cyl -m /path/to/metadata -d /path/to/data` results in
 the following operations:
