+++ lib/traject/extract_work_type.rb
@@ -17,9 +17,12 @@ WORK_TYPE_MAP = {
 }.freeze
 
 module ExtractWorkType
-  # Transfor leader field 006 into a LOC Resource Type URI
+  # Transfer leader field 006 into a LOC Resource Type URI
   def extract_work_type
     lambda do |record, accumulator|
+      # example:
+      #   LEADER 01488njm a2200385 a 4500
+      # we want "j" from that leader field
       accumulator << WORK_TYPE_MAP[record.leader.slice(6)]
     end
   end
